Fix issue that can lead to i2c clock frequency exceeding spec (#497)
The i2c hardware doesn't support much precision in selecting duty cycles
- we only have six bits to use to determine the duty cycle, and then we
have to come up with a 'clock divider' that hits the target speed as
closely as possible. In cases where that math doesn't divide precisely,
we need to select the closest clock divider that doesn't exceed the max
speed. However, the current implementation just selects the clock
divider that minimizes the difference between target frequency and
actual frequency, without regard for which side of the target frequency
that error lands on.
This change removes clock speeds that would exceed the spec from
consideration for best fit so we don't violate the spec.